[!U] Two different UV sets were created for the buildings.
Each building has Cylindrical UV mapping for the sides and the top / bottom is a UV Projection. These can be separated in the material with Vertex Colors. The sides have a Red vertex color and the top / bottom have a green vertex color. The second building has more subdivisions in the geometry and the UVs are scaled. Base UV resolution for both is 2048. UV textures have also been added.